+++ src/test/java/org/morphix/reflection/MethodTypeTest.java
@@ -14,6 +14,7 @@ package org.morphix.reflection;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -187,6 +188,7 @@ class MethodTypeTest {
 	}
 
 	public static class C {
+
 		public String getC() {
 			return null;
 		}
@@ -199,8 +201,12 @@ class MethodTypeTest {
 			return null;
 		}
 
-		public String getY(@SuppressWarnings("unused") final int x) {
+		public String getY(@SuppressWarnings("unused") final int x) { // NOSONAR need it for test
 			return null;
+		}
+
+		public void popX(@SuppressWarnings("unused") final String x) {
+			// empty
 		}
 	}
 
@@ -249,6 +255,15 @@ class MethodTypeTest {
 	}
 
 	@Test
+	void shouldReturnFalseForSetterPredicateOnNonSetter() throws Exception {
+		Method method = C.class.getDeclaredMethod("popX", String.class);
+
+		boolean isSetter = MethodType.SETTER.getPredicate().test(method);
+
+		assertFalse(isSetter);
+	}
+
+	@Test
 	void shouldReturnFalseForGetterPredicateOnNonGetterWithParameters() throws Exception {
 		Method method = C.class.getDeclaredMethod("getY", int.class);
 
@@ -257,4 +272,41 @@ class MethodTypeTest {
 		assertFalse(isGetter);
 	}
 
+	public static record D(String x) {
+		// empty
+	}
+
+	@Test
+	void shouldReturnTheFieldNameForRecordGetter() throws Exception {
+		Method method = D.class.getDeclaredMethod("x");
+
+		String fieldName = MethodType.GETTER.getFieldName(method);
+
+		assertThat(fieldName, equalTo("x"));
+	}
+
+	@Test
+	void shouldReturnNullForRecordSetter() throws Exception {
+		Method method = D.class.getDeclaredMethod("x");
+
+		String fieldName = MethodType.SETTER.getFieldName(method);
+
+		assertThat(fieldName, nullValue());
+	}
+
+	public static record E(String x) {
+
+		String x(final int x) {
+			return "" + this.x + x;
+		}
+	}
+
+	@Test
+	void shouldReturnNullForRenamedRecordGetter() throws Exception {
+		Method method = E.class.getDeclaredMethod("x", int.class);
+
+		String fieldName = MethodType.GETTER.getFieldName(method);
+
+		assertThat(fieldName, nullValue());
+	}
 }